refactoring(cli) - update after rebase and ensure we deploy contracts with stateTreeDepth
