+++ cli/tests/e2e.test.ts
@@ -1,4 +1,5 @@
 import {
+    checkVerifyingKeys,
     deploy,
     deployPoll,
     deployVkRegistryContract, 
@@ -21,16 +22,15 @@ import {
     coordinatorPrivKey, 
     coordinatorPubKey
 } from "./utilities/constants"
-import {
-    homedir
-} from "os"
-import { cleanSubsidy } from "./utilities/utils"
+import { homedir } from "os"
+import { cleanVanilla } from "./utilities/utils"
 import { isArm } from "maci-circuits"
 
-describe("Subsidy tests", function() {
+describe("e2e tests", function() {
     const useWasm = isArm() 
     this.timeout(900000)
   
+    // before all tests we deploy the vk registry contract and set the verifying keys
     before(async () => {
         // we deploy the vk registry contract
         await deployVkRegistryContract(true)
@@ -44,17 +44,18 @@ describe("Subsidy tests", function() {
             "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
             "./zkeys/TallyVotes_10-1-2_test.0.zkey",   
             undefined,
-            "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+            undefined,
             true
         )
     })
+    
     describe("test1", () => {
         after(async () => {
-            cleanSubsidy()
+            cleanVanilla()
         })
         before(async () => {
             // deploy the smart contracts
-            await deploy(undefined, undefined, undefined, undefined, true)
+            await deploy(STATE_TREE_DEPTH, undefined, undefined, undefined, undefined, true)
             // deploy a poll contract
             await deployPoll(
                 90,
@@ -111,18 +112,18 @@ describe("Subsidy tests", function() {
                 "./zkeys/TallyVotes_10-1-2_test.0.zkey",
                 "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
                 0,
-                "./subsidy.json",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+                undefined,
+                undefined,
                 `${homedir()}/rapidsnark/build/prover`,
                 "./zkeys/ProcessMessages_10-2-1-2_test",
                 "./zkeys/TallyVotes_10-1-2_test",
-                "./zkeys/SubsidyPerBatch_10-1-2_test",
+                undefined,
                 coordinatorPrivKey,
                 undefined,
                 undefined,
                 "./zkeys/ProcessMessages_10-2-1-2_test.wasm",
                 "./zkeys/TallyVotes_10-1-2_test.wasm",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.wasm",
+                undefined,
                 useWasm,
                 true
             )
@@ -143,18 +144,18 @@ describe("Subsidy tests", function() {
                 undefined,
                 undefined,
                 undefined,
-                "./subsidy.json",
+                undefined,
                 true
             )
         })
     })
 
     describe("test2", () => {
-        after (() => cleanSubsidy())
+        after (() => cleanVanilla())
 
         before(async () => {
             // deploy the smart contracts
-            await deploy(undefined, undefined, undefined, undefined, true)
+            await deploy(STATE_TREE_DEPTH, undefined, undefined, undefined, undefined, true)
             // deploy a poll contract
             await deployPoll(
                 90,
@@ -302,18 +303,18 @@ describe("Subsidy tests", function() {
                 "./zkeys/TallyVotes_10-1-2_test.0.zkey",
                 "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
                 0,
-                "./subsidy.json",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+                undefined,
+                undefined,
                 `${homedir()}/rapidsnark/build/prover`,
                 "./zkeys/ProcessMessages_10-2-1-2_test",
                 "./zkeys/TallyVotes_10-1-2_test",
-                "./zkeys/SubsidyPerBatch_10-1-2_test",
+                undefined,
                 coordinatorPrivKey,
                 undefined,
                 undefined,
                 "./zkeys/ProcessMessages_10-2-1-2_test.wasm",
                 "./zkeys/TallyVotes_10-1-2_test.wasm",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.wasm",
+                undefined,
                 useWasm,
                 true
             )
@@ -334,18 +335,18 @@ describe("Subsidy tests", function() {
                 undefined,
                 undefined,
                 undefined,
-                "./subsidy.json",
+                undefined,
                 true
             )
         })
     })
 
     describe("test3", () => {
-        after (() => cleanSubsidy())
+        after (() => cleanVanilla())
 
         before(async () => {
             // deploy the smart contracts
-            await deploy(undefined, undefined, undefined, undefined, true)
+            await deploy(STATE_TREE_DEPTH, undefined, undefined, undefined, undefined, true)
             // deploy a poll contract
             await deployPoll(
                 90,
@@ -468,18 +469,18 @@ describe("Subsidy tests", function() {
                 "./zkeys/TallyVotes_10-1-2_test.0.zkey",
                 "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
                 0,
-                "./subsidy.json",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+                undefined,
+                undefined,
                 `${homedir()}/rapidsnark/build/prover`,
                 "./zkeys/ProcessMessages_10-2-1-2_test",
                 "./zkeys/TallyVotes_10-1-2_test",
-                "./zkeys/SubsidyPerBatch_10-1-2_test",
+                undefined,
                 coordinatorPrivKey,
                 undefined,
                 undefined,
                 "./zkeys/ProcessMessages_10-2-1-2_test.wasm",
                 "./zkeys/TallyVotes_10-1-2_test.wasm",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.wasm",
+                undefined,
                 useWasm,
                 true
             )
@@ -500,7 +501,7 @@ describe("Subsidy tests", function() {
                 undefined,
                 undefined,
                 undefined,
-                "./subsidy.json",
+                undefined,
                 true
             )
         })
@@ -508,11 +509,11 @@ describe("Subsidy tests", function() {
     })
 
     describe("test4", () => {
-        after (() => cleanSubsidy())
+        after (() => cleanVanilla())
         
         before(async () => {
             // deploy the smart contracts
-            await deploy(undefined, undefined, undefined, undefined, true)
+            await deploy(STATE_TREE_DEPTH, undefined, undefined, undefined, undefined, true)
             // deploy a poll contract
             await deployPoll(
                 90,
@@ -579,18 +580,18 @@ describe("Subsidy tests", function() {
                 "./zkeys/TallyVotes_10-1-2_test.0.zkey",
                 "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
                 0,
-                "./subsidy.json",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+                undefined,
+                undefined,
                 `${homedir()}/rapidsnark/build/prover`,
                 "./zkeys/ProcessMessages_10-2-1-2_test",
                 "./zkeys/TallyVotes_10-1-2_test",
-                "./zkeys/SubsidyPerBatch_10-1-2_test",
+                undefined,
                 coordinatorPrivKey,
                 undefined,
                 undefined,
                 "./zkeys/ProcessMessages_10-2-1-2_test.wasm",
                 "./zkeys/TallyVotes_10-1-2_test.wasm",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.wasm",
+                undefined,
                 useWasm,
                 true
             )
@@ -612,17 +613,17 @@ describe("Subsidy tests", function() {
                 undefined,
                 undefined,
                 undefined,
-                "./subsidy.json",
+                undefined,
                 true
             )
         })
     })
 
     describe("test5", () => {
-        after (() => cleanSubsidy())
+        after (() => cleanVanilla())
         before(async () => {
             // deploy the smart contracts
-            await deploy(undefined, undefined, undefined, undefined, true)
+            await deploy(STATE_TREE_DEPTH, undefined, undefined, undefined, undefined, true)
             // deploy a poll contract
             await deployPoll(
                 90,
@@ -746,18 +747,18 @@ describe("Subsidy tests", function() {
                 "./zkeys/TallyVotes_10-1-2_test.0.zkey",
                 "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
                 0,
-                "./subsidy.json",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+                undefined,
+                undefined,
                 `${homedir()}/rapidsnark/build/prover`,
                 "./zkeys/ProcessMessages_10-2-1-2_test",
                 "./zkeys/TallyVotes_10-1-2_test",
-                "./zkeys/SubsidyPerBatch_10-1-2_test",
+                undefined,
                 coordinatorPrivKey,
                 undefined,
                 undefined,
                 "./zkeys/ProcessMessages_10-2-1-2_test.wasm",
                 "./zkeys/TallyVotes_10-1-2_test.wasm",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.wasm",
+                undefined,
                 useWasm,
                 true
             )
@@ -779,18 +780,18 @@ describe("Subsidy tests", function() {
                 undefined,
                 undefined,
                 undefined,
-                "./subsidy.json",
+                undefined,
                 true
             )
         })
     })
 
     describe("test6", () => {
-        after (() => cleanSubsidy())
+        after (() => cleanVanilla())
 
         before(async () => {
             // deploy the smart contracts
-            await deploy(undefined, undefined, undefined, undefined, true)
+            await deploy(STATE_TREE_DEPTH, undefined, undefined, undefined, undefined, true)
             // deploy a poll contract
             await deployPoll(
                 90,
@@ -933,18 +934,18 @@ describe("Subsidy tests", function() {
                 "./zkeys/TallyVotes_10-1-2_test.0.zkey",
                 "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
                 0,
-                "./subsidy.json",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+                undefined,
+                undefined,
                 `${homedir()}/rapidsnark/build/prover`,
                 "./zkeys/ProcessMessages_10-2-1-2_test",
                 "./zkeys/TallyVotes_10-1-2_test",
-                "./zkeys/SubsidyPerBatch_10-1-2_test",
+                undefined,
                 coordinatorPrivKey,
                 undefined,
                 undefined,
                 "./zkeys/ProcessMessages_10-2-1-2_test.wasm",
                 "./zkeys/TallyVotes_10-1-2_test.wasm",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.wasm",
+                undefined,
                 useWasm,
                 true
             )
@@ -966,122 +967,38 @@ describe("Subsidy tests", function() {
                 undefined,
                 undefined,
                 undefined,
-                "./subsidy.json",
+                undefined,
                 true
             )
         })
     })
 
+    describe("checkKeys", () => {
+        before(async () => {
+            // deploy maci as we need the address
+            await deploy(STATE_TREE_DEPTH, undefined, undefined, undefined, undefined, true)
+        })
+        it("should check if the verifying keys have been set correctly", async () => {
+            await checkVerifyingKeys(
+                STATE_TREE_DEPTH,
+                INT_STATE_TREE_DEPTH,
+                MSG_TREE_DEPTH,
+                VOTE_OPTION_TREE_DEPTH,
+                MSG_BATCH_DEPTH,
+                "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
+                "./zkeys/TallyVotes_10-1-2_test.0.zkey",
+                undefined,
+                true,
+            )
+        })
+    })
 
-    describe("test7", () => {
-        after(() => { cleanSubsidy() })
-
-        const publishArgs = [
-            {
-                quiet: true,
-                pubkey: "macipk.3e7bb2d7f0a1b7e980f1b6f363d1e3b7a12b9ae354c2cd60a9cfa9fd12917391",
-                privateKey: "macisk.fd7aa614ec4a82716ffc219c24fd7e7b52a2b63b5afb17e81c22fe21515539c",
-                stateIndex: 6,
-                voteOptionIndex: 5,
-                nonce: 1,
-                newVoteWeight: 6,
-                pollId: 0
-            },
-            {
-                quiet: true,
-                pubkey: "macipk.3e7bb2d7f0a1b7e980f1b6f363d1e3b7a12b9ae354c2cd60a9cfa9fd12917391",
-                privateKey: "macisk.fd7aa614ec4a82716ffc219c24fd7e7b52a2b63b5afb17e81c22fe21515539c",
-                stateIndex: 5,
-                voteOptionIndex: 5,
-                nonce: 2,
-                newVoteWeight: 5,
-                pollId: 0
-            },
-            {
-                quiet: true,
-                pubkey: "macipk.3e7bb2d7f0a1b7e980f1b6f363d1e3b7a12b9ae354c2cd60a9cfa9fd12917391",
-                privateKey: "macisk.fd7aa614ec4a82716ffc219c24fd7e7b52a2b63b5afb17e81c22fe21515539c",
-                stateIndex: 5,
-                voteOptionIndex: 4,
-                nonce: 1,
-                pollId: 0,
-                newVoteWeight: 5
-            },
-            {
-                quiet: true,
-                pubkey: "macipk.3e7bb2d7f0a1b7e980f1b6f363d1e3b7a12b9ae354c2cd60a9cfa9fd12917391",
-                privateKey: "macisk.fd7aa614ec4a82716ffc219c24fd7e7b52a2b63b5afb17e81c22fe21515539c",
-                stateIndex: 4,
-                voteOptionIndex: 4,
-                nonce: 2,
-                pollId: 0,
-                newVoteWeight: 4
-            },
-            {
-                quiet: true,
-                pubkey: "macipk.3e7bb2d7f0a1b7e980f1b6f363d1e3b7a12b9ae354c2cd60a9cfa9fd12917391",
-                privateKey: "macisk.fd7aa614ec4a82716ffc219c24fd7e7b52a2b63b5afb17e81c22fe21515539c",
-                stateIndex: 4,
-                voteOptionIndex: 3,
-                nonce: 1,
-                pollId: 0,
-                newVoteWeight: 4
-            }, 
-            {
-                quiet: true,
-                pubkey: "macipk.3e7bb2d7f0a1b7e980f1b6f363d1e3b7a12b9ae354c2cd60a9cfa9fd12917391",
-                privateKey: "macisk.fd7aa614ec4a82716ffc219c24fd7e7b52a2b63b5afb17e81c22fe21515539c",
-                stateIndex: 3,
-                voteOptionIndex: 3,
-                nonce: 2,
-                pollId: 0,
-                newVoteWeight: 3
-            }, 
-            {
-                quiet: true,
-                pubkey: "macipk.3e7bb2d7f0a1b7e980f1b6f363d1e3b7a12b9ae354c2cd60a9cfa9fd12917391",
-                privateKey: "macisk.fd7aa614ec4a82716ffc219c24fd7e7b52a2b63b5afb17e81c22fe21515539c",
-                stateIndex: 2,
-                voteOptionIndex: 2,
-                nonce: 2,
-                pollId: 0,
-                newVoteWeight: 2
-            }, 
-            {
-                quiet: true,
-                pubkey: "macipk.3e7bb2d7f0a1b7e980f1b6f363d1e3b7a12b9ae354c2cd60a9cfa9fd12917391",
-                privateKey: "macisk.fd7aa614ec4a82716ffc219c24fd7e7b52a2b63b5afb17e81c22fe21515539c",
-                stateIndex: 2,
-                voteOptionIndex: 1,
-                nonce: 1,
-                pollId: 0,
-                newVoteWeight: 2
-            }, 
-            {
-                quiet: true,
-                pubkey: "macipk.3e7bb2d7f0a1b7e980f1b6f363d1e3b7a12b9ae354c2cd60a9cfa9fd12917391",
-                privateKey: "macisk.fd7aa614ec4a82716ffc219c24fd7e7b52a2b63b5afb17e81c22fe21515539c",
-                stateIndex: 1,
-                voteOptionIndex: 1,
-                nonce: 2,
-                pollId: 0,
-                newVoteWeight: 1
-            },
-            {
-                quiet: true,
-                pubkey: "macipk.3e7bb2d7f0a1b7e980f1b6f363d1e3b7a12b9ae354c2cd60a9cfa9fd12917391",
-                privateKey: "macisk.fd7aa614ec4a82716ffc219c24fd7e7b52a2b63b5afb17e81c22fe21515539c",
-                stateIndex: 1,
-                voteOptionIndex: 0,
-                nonce: 1,
-                pollId: 0,
-                newVoteWeight: 1
-            }, 
-        ]
+    describe("keyChange", () => {
+        after (() => cleanVanilla())
 
         before(async () => {
             // deploy the smart contracts
-            await deploy(undefined, undefined, undefined, undefined, true)
+            await deploy(STATE_TREE_DEPTH, undefined, undefined, undefined, undefined, true)
             // deploy a poll contract
             await deployPoll(
                 90,
@@ -1095,29 +1012,55 @@ describe("Subsidy tests", function() {
                 undefined,
                 true,
             )
-        })
-
-        it("should signup an user", async () => {
-            for (let i = 0; i < 6; i++) await signup(
-                "macipk.3e7bb2d7f0a1b7e980f1b6f363d1e3b7a12b9ae354c2cd60a9cfa9fd12917391",
+            // signup a user
+            await signup(
+                "macipk.b8590fdba5e9cde5606dad5db384be4d253d0a2064d1e03f9600ee021a7ebe16", 
+                undefined, 
+                undefined, 
+                undefined, 
+                true
+            )
+            // publish messages
+            await publish(
+                "macipk.b8590fdba5e9cde5606dad5db384be4d253d0a2064d1e03f9600ee021a7ebe16",
+                1,
+                0,
+                1,
+                0,
+                9,
                 undefined,
                 undefined,
-                undefined,
+                "macisk.2ae4f199bf3925a2407f7c775c9261f351ab861d8e9ecbb84622bdd3f6d41b08",
                 true
             )
         })
 
-        it("should publish all messages", async () => {
-            for (const arg of publishArgs) await publish(
-                arg.pubkey,
-                arg.stateIndex,
-                arg.voteOptionIndex,
-                arg.nonce,
-                arg.pollId,
-                arg.newVoteWeight,
+        it("should publish a message to change the user maci key", async () => {
+            await publish(
+                "macipk.b42b0da48010682d8c781d403f6b83db00c5e0970094ef3618393e7a3262c320",
+                1,
+                0,
+                1,
+                0,
+                9,
                 undefined,
                 undefined,
-                arg.privateKey,
+                "macisk.2ae4f199bf3925a2407f7c775c9261f351ab861d8e9ecbb84622bdd3f6d41b08",
+                true
+            )
+        })
+
+        it("should vote for a different option using the new key", async () => {
+            await publish(
+                "macipk.b42b0da48010682d8c781d403f6b83db00c5e0970094ef3618393e7a3262c320",
+                1,
+                1,
+                2,
+                0,
+                9,
+                undefined,
+                undefined,
+                "macisk.220b09bca39ddc56deaaecddcdf616529cd2ed3eeda2354795515f17894e1c65",
                 true
             )
         })
@@ -1148,18 +1091,18 @@ describe("Subsidy tests", function() {
                 "./zkeys/TallyVotes_10-1-2_test.0.zkey",
                 "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
                 0,
-                "./subsidy.json",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+                undefined,
+                undefined,
                 `${homedir()}/rapidsnark/build/prover`,
                 "./zkeys/ProcessMessages_10-2-1-2_test",
                 "./zkeys/TallyVotes_10-1-2_test",
-                "./zkeys/SubsidyPerBatch_10-1-2_test",
+                undefined,
                 coordinatorPrivKey,
                 undefined,
                 undefined,
                 "./zkeys/ProcessMessages_10-2-1-2_test.wasm",
                 "./zkeys/TallyVotes_10-1-2_test.wasm",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.wasm",
+                undefined,
                 useWasm,
                 true
             )
@@ -1181,17 +1124,17 @@ describe("Subsidy tests", function() {
                 undefined,
                 undefined,
                 undefined,
-                "./subsidy.json",
+                undefined,
                 true
             )
         })
     })
 
     describe("multiplePolls1", () => {
-        after (() => cleanSubsidy())
+        after (() => cleanVanilla())
         before(async () => {
             // deploy the smart contracts
-            await deploy(undefined, undefined, undefined, undefined, true)
+            await deploy(STATE_TREE_DEPTH, undefined, undefined, undefined, undefined, true)
             // deploy a poll contract
             await deployPoll(
                 90,
@@ -1237,18 +1180,18 @@ describe("Subsidy tests", function() {
                 "./zkeys/TallyVotes_10-1-2_test.0.zkey",
                 "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
                 0,
-                "./subsidy.json",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+                undefined,
+                undefined,
                 `${homedir()}/rapidsnark/build/prover`,
                 "./zkeys/ProcessMessages_10-2-1-2_test",
                 "./zkeys/TallyVotes_10-1-2_test",
-                "./zkeys/SubsidyPerBatch_10-1-2_test",
+                undefined,
                 coordinatorPrivKey,
                 undefined,
                 undefined,
                 "./zkeys/ProcessMessages_10-2-1-2_test.wasm",
                 "./zkeys/TallyVotes_10-1-2_test.wasm",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.wasm",
+                undefined,
                 useWasm,
                 true
             )
@@ -1268,10 +1211,10 @@ describe("Subsidy tests", function() {
                 undefined,
                 undefined,
                 undefined,
-                "./subsidy.json",
+                undefined,
                 true 
             )
-            cleanSubsidy()
+            cleanVanilla()
         })
 
         it("should deploy a new poll", async () => {
@@ -1312,18 +1255,18 @@ describe("Subsidy tests", function() {
                 "./zkeys/TallyVotes_10-1-2_test.0.zkey",
                 "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
                 1,
-                "./subsidy.json",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+                undefined,
+                undefined,
                 `${homedir()}/rapidsnark/build/prover`,
                 "./zkeys/ProcessMessages_10-2-1-2_test",
                 "./zkeys/TallyVotes_10-1-2_test",
-                "./zkeys/SubsidyPerBatch_10-1-2_test",
+                undefined,
                 coordinatorPrivKey,
                 undefined,
                 undefined,
                 "./zkeys/ProcessMessages_10-2-1-2_test.wasm",
                 "./zkeys/TallyVotes_10-1-2_test.wasm",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.wasm",
+                undefined,
                 useWasm,
                 true
             )
@@ -1336,7 +1279,7 @@ describe("Subsidy tests", function() {
                 undefined,
                 true
             )
-            await verify('1', './tally.json', undefined, undefined, undefined, "./subsidy.json", true)
+            await verify('1', './tally.json', undefined, undefined, undefined, undefined, true)
         })
     })
 
@@ -1351,11 +1294,11 @@ describe("Subsidy tests", function() {
             "macipk.d354c22572a0b53ced8561f9ab16abe0b8f7f98c2133757e93cdefd45fe1b192"
         ]
 
-        after(() => { cleanSubsidy() })
+        after(() => { cleanVanilla() })
 
         before(async () => {
             // deploy the smart contracts
-            await deploy(undefined, undefined, undefined, undefined, true)
+            await deploy(STATE_TREE_DEPTH, undefined, undefined, undefined, undefined, true)
         })
 
         it("should run the first poll", async () => {
@@ -1404,18 +1347,18 @@ describe("Subsidy tests", function() {
                 "./zkeys/TallyVotes_10-1-2_test.0.zkey",
                 "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
                 0,
-                "./subsidy.json",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+                undefined,
+                undefined,
                 `${homedir()}/rapidsnark/build/prover`,
                 "./zkeys/ProcessMessages_10-2-1-2_test",
                 "./zkeys/TallyVotes_10-1-2_test",
-                "./zkeys/SubsidyPerBatch_10-1-2_test",
+                undefined,
                 coordinatorPrivKey,
                 undefined,
                 undefined,
                 "./zkeys/ProcessMessages_10-2-1-2_test.wasm",
                 "./zkeys/TallyVotes_10-1-2_test.wasm",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.wasm",
+                undefined,
                 useWasm,
                 true
             )
@@ -1435,10 +1378,10 @@ describe("Subsidy tests", function() {
                 undefined,
                 undefined,
                 undefined,
-                "./subsidy.json",
+                undefined,
                 true 
             )
-            cleanSubsidy()
+            cleanVanilla()
         })
 
         it("should deploy two more polls", async () => {
@@ -1562,18 +1505,18 @@ describe("Subsidy tests", function() {
                 "./zkeys/TallyVotes_10-1-2_test.0.zkey",
                 "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
                 1,
-                "./subsidy.json",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+                undefined,
+                undefined,
                 `${homedir()}/rapidsnark/build/prover`,
                 "./zkeys/ProcessMessages_10-2-1-2_test",
                 "./zkeys/TallyVotes_10-1-2_test",
-                "./zkeys/SubsidyPerBatch_10-1-2_test",
+                undefined,
                 coordinatorPrivKey,
                 undefined,
                 undefined,
                 "./zkeys/ProcessMessages_10-2-1-2_test.wasm",
                 "./zkeys/TallyVotes_10-1-2_test.wasm",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.wasm",
+                undefined,
                 useWasm,
                 true
             )
@@ -1592,10 +1535,10 @@ describe("Subsidy tests", function() {
                 undefined,
                 undefined,
                 undefined,
-                "./subsidy.json",
+                undefined,
                 true
             )
-            cleanSubsidy()
+            cleanVanilla()
         })
 
         it("should complete the third poll", async () => {
@@ -1607,18 +1550,18 @@ describe("Subsidy tests", function() {
                 "./zkeys/TallyVotes_10-1-2_test.0.zkey",
                 "./zkeys/ProcessMessages_10-2-1-2_test.0.zkey",
                 2,
-                "./subsidy.json",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.0.zkey",
+                undefined,
+                undefined,
                 `${homedir()}/rapidsnark/build/prover`,
                 "./zkeys/ProcessMessages_10-2-1-2_test",
                 "./zkeys/TallyVotes_10-1-2_test",
-                "./zkeys/SubsidyPerBatch_10-1-2_test",
+                undefined,
                 coordinatorPrivKey,
                 undefined,
                 undefined,
                 "./zkeys/ProcessMessages_10-2-1-2_test.wasm",
                 "./zkeys/TallyVotes_10-1-2_test.wasm",
-                "./zkeys/SubsidyPerBatch_10-1-2_test.wasm",
+                undefined,
                 useWasm,
                 true
             )
@@ -1637,10 +1580,9 @@ describe("Subsidy tests", function() {
                 undefined,
                 undefined,
                 undefined,
-                "./subsidy.json",
+                undefined,
                 true
             )
         })
     })
-
 })
